pyproject.toml
# =============================================================================
# finance-copilot — pyproject.toml
# Central project metadata and tool configuration.
# Replaces setup.py, setup.cfg, and disparate tool config files.
# =============================================================================

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-copilot"
version = "0.1.0"
description = "Corporate Finance Autopilot — agentic multi-step financial analysis pipeline"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "José Ramón Morera" }]

# Runtime dependencies are pinned in requirements.txt for Docker reproducibility.
# This section mirrors them for editable installs.
dependencies = [
    # --- Web framework ---
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    # --- Config / secrets ---
    "pydantic-settings>=2.2",
    # --- AI / LLM orchestration ---
    "langchain>=0.2",
    "langchain-openai>=0.1",
    "langchain-google-genai>=1.0",
    "langgraph>=0.1",          # multi-agent graph orchestration
    # --- Financial data ---
    "yfinance>=0.2",
    "edgartools>=3.0",
    "pandas>=2.2",
    "numpy>=1.26",
    # --- Data validation ---
    "pydantic>=2.6",
    # --- Database ---
    "sqlmodel>=0.0.16",
    "psycopg[binary]>=3.1",
    # --- HTTP / scraping ---
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    # --- Outputs ---
    "python-pptx>=0.6",
    "reportlab>=4.1",
    "jinja2>=3.1",
    "openpyxl>=3.1",
    # --- Observability ---
    "structlog>=24.1",
    "opentelemetry-sdk>=1.24",
    "langfuse>=2.0",
    "chromadb>=0.6.3",
    "Pillow>=10.0,<11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",         # AsyncClient for FastAPI tests
    "ruff>=0.4",
    "mypy>=1.9",
    "pre-commit>=3.7",
    "types-beautifulsoup4",
    "pandas-stubs",
]

# =============================================================================
# Ruff — linter + formatter (replaces flake8, isort, black)
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ANN", # annotations (light subset)
]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.format]
quote-style = "double"

# =============================================================================
# Mypy — static type checker
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

# =============================================================================
# Pytest
# =============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
